third_party/WebKit/LayoutTests/media/video-controls-no-scripting.html

<!DOCTYPE html>
<title>Tests that the built-in controls are always enabled when JavaScript is disabled.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<iframe sandbox="allow-same-origin" src="resources/video-controls-no-scripting-iframe.html" style="width: 400px; height: 320px;"></iframe>
<script>
async_test(function(t) {
    var iframe = document.querySelector("iframe");

    iframe.onload = t.step_func_done(function() {
        var video = iframe.contentDocument.querySelector("video");
        assert_not_equals(internals.shadowRoot(video), null);
        var controls = mediaControlsButton(video, "panel");
        assert_equals(getComputedStyle(controls).opacity, "1");
    });
});
</script>
